separate routers for airports and airspaces
